+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 * Hide the constructors of `AnyGet/AnyPut/AnyDelete/AnyScan/...` because the static creator method `XXX.of(...)` is preferred.
 
+* Rename `NamedCQL` to `ParsedCql`.
+
 * Improvements and bug fix.